Handle zod errors and return 400 instead of 500
By default express will catch all errors and return 500.
In case there's a ZodError, it means that the request validation failed, which means the client sent wrong data.
When client sends wrong data, 400 should be returned because it means that the error is on client side and 500 means the error is on the server side.